JS Game Boy Emulator: Add gamepad support
Add HTML5 gamepad support to the javascript gameboy emulator.
When w3c "standard" mapping is available for the gamepad, buttons
will map to the expected locations Otherwise the first 4 buttons
map to A/B/Select/Start.

Gamepad support co-exists and does not interfere with keyboard and
on-screen controls. Polling only starts if a button is pressed on
gamepad, and will stop if the gamepad disconnects.

Sound will also start when the gamepad polling is started, similar
to keyboard input.