Add a "Save As" feature
Implements a save as feature requested by #103. Allows users to
save a copy of a modified or unmodified project without reloading.

GB Studio now displays a "Save As..." option in the File dropdown.
When clicked a dialog opens that allows the user to name a new
project. Once the user clicks save the project is copied to
the new location requested and the modified state is updated
if it was set. Also includes an update to the Chinese language
pack to include "Save As".