Custom Events 2
The first version of Custom Events worked by duplicating the script
inside the Event on each of its instances. This meant that scripts
could grow too big without users even noticing.

This update makes Custom Events single scripts in memory and each
call to them jumps the script execution to the memory position and
bank where the script is located. Before that, the parameter values
are copied into "local" variables. Once the script is executed the
local variable values are copied back to the parameter ones,

For actors, a new internal C script has been introduced that allows
to set the active actor from a variable value.