GBT speed conversion, better A0x deprication, Fix ch3 volume only bug
To disable default speed conversion, add -speed to end of args.
A0x sends nop effect if with note, or ignored if effect only, intended for faked previews in tracker.
Ch3 volume only was 1 bit to the right, fixed in mod2gbt.
Start of support for ch1-2 frequency without trigger