Revert GBT to always process ch1-2 instrument now that fixes are done
At a previous time, This was changed to check if an effect needed to update registers, as effects like pan would not need to, but would cause a note to play. With split trigger and no trigger paths, along with channel instrument prediction, this can be reverted, retaining more compatibility.