Add x and y checkboxes to camera shake command
NOTE: This changes the format of the "CameraShake" command!

This means the new gbdk game engine code (after this commit) will no
longer be feature-compatible with the front-end code (before this
commit). The opposite is also true.

I don't know how often command formats are changed upstream, but I hope
this feature is found useful enough to change the command. I had a lot
of fun working on implementing it!