Attach script to button overrides default input
In 1.2 using "Attach script to button" disabled the default behavior for the button. So, attaching a script to A will disable the interaction with actors.

In 2.0 when attaching, the engine will try to run the two scripts in the background which lead to some unexpected issues like the player moving even if the directional buttons had an script attached ( #452 , #502 ).

This patch restores the 1.2 behavior.