Add "Actor: Face Player"
Adds a new event, "Actor: Face Player".

This changes an actor's direction to face the player. By defaul, it determines whether the X or Y direction has a greater magnitude of difference between the player and the specified actor and sets the actor's direction to the greater. You can also specify if you only want the actor to face left/right ("Horizontal Only") and up/down ("Vertical Only"). You can finally invert the direction the actor faces so it faces away from the player.