Fix problem with centering scenes
- A problem with the comparison of "onlyMatchingScene" caused every
change to a given scene to zoom to the center of said scene, making
it literally impossible to paint big scenes with palettes/collisions.
- Changed said comparison to an string match of ids, since objects
cannot be compared unless their pointers are equal.
- Additionally, added a preventDefault() call on middle-click panning
of the map, as middle click on windows causes a scroll arrow, which
literally fought against the panning action.